fix: move download function to utils
