+++ src/ui/pages/DownloadBackup/useDownloadBackup.ts
@@ -5,6 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { PasswordFormFields } from "@src/types";
 import { downloadBackup } from "@src/ui/ducks/backup";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
+import { downloadFile } from "@src/util/browser";
 import { formatDate } from "@src/util/date";
 
 export interface IUseDownloadBackupData {
@@ -41,22 +42,10 @@ export const useDownloadBackup = (): IUseDownloadBackupData => {
     navigate(-1);
   }, [navigate]);
 
-  const downloadFile = useCallback((content: string) => {
-    const element = document.createElement("a");
-    element.style.display = "none";
-    element.setAttribute("href", content);
-    element.setAttribute("download", `ck-backup-${formatDate(new Date())}.json`);
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-
-    return Promise.resolve();
-  }, []);
-
   const onSubmit = useCallback(
     (data: DownloadBackupFields) => {
       dispatch(downloadBackup(data.password))
-        .then((content: string) => downloadFile(content))
+        .then((content: string) => downloadFile(content, `ck-backup-${formatDate(new Date())}.json`))
         .then(() => onGoBack())
         .catch((error: Error) => setError("password", { type: "submit", message: error.message }));
     },
